pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "kaleido>=1.1.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "razorpay>=2.0.0",
    "reportlab>=4.4.4",
    "streamlit>=1.50.0",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
    "yfinance>=0.2.66",
]
